requirements.txt
streamlit>=1.32.0
pm4py>=2.7.7
google-generativeai>=0.3.2
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
matplotlib>=3.8.0
graphviz>=0.20.1
pydotplus>=2.0.2
pathlib>=1.0.1
